Remove compile warnings on windows
The code emitted too many warnings making it impossible
to see if there was any serious warnings being reported

Most of the changes are due to return values doing conversions
that would cause some sort of data loss.

Reviewed-on: http://review.couchbase.org/32736
Tested-by: Trond Norbye <trond.norbye@gmail.com>

+++ daemon/cache.h
@@ -53,9 +53,9 @@ typedef struct {
     /** The size of each element in this cache */
     size_t bufsize;
     /** The capacity of the list of elements */
-    int freetotal;
+    size_t freetotal;
     /** The current number of free elements */
-    int freecurr;
+    size_t freecurr;
     /** The constructor to be called each time we allocate more memory */
     cache_constructor_t* constructor;
     /** The destructor to be called each time before we release memory */
